Make User Profile

+++ src/controllers/videoController.js
@@ -1,5 +1,6 @@
 import e from 'express';
 import Video from '../models/Video';
+import User from '../models/User';
 
 export const home = async (req, res) => {
   try {
@@ -12,7 +13,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate('owner');
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
   }
@@ -21,14 +22,23 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   return res.render('edit', { pageTitle: `Editing`, video });
 };
 
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const video = await Video.exists({ _id: id });
   const { title, description, hashtags } = req.body;
@@ -36,6 +46,11 @@ export const postEdit = async (req, res) => {
   if (!video) {
     return res.render('404', { pageTitle: 'Video not found.' });
   }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
+
   await Video.findByIdAndUpdate(id, {
     title,
     description,
@@ -50,14 +65,26 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
 
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+
+    const user = await User.findById(_id);
+
+    user.videos.push(newVideo._id);
+    user.save();
+
     return res.redirect('/');
   } catch (error) {
     return res.status(400).render('upload', {
@@ -69,6 +96,20 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+
+  const video = await Video.findById(id);
+
+  if (!video) {
+    return res.status(404).render('404', { pageTitle: 'Video not found.' });
+  }
+
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
+
   await Video.findByIdAndDelete(id);
   return res.redirect('/');
 };
